python/python-library-recipe/requirements.txt
python-dateutil==2.8.2
pytz==2022.5
PyYAML==6.0
